feat: implement modularized UI templates and state management for settings page
